+++ rocAL/rocAL/source/api/rocal_api_meta_data.cpp
@@ -390,8 +390,9 @@ ROCAL_API_CALL rocalGetOneHotImageLabels(RocalContext p_context, void* buf, int 
     }
 }
 
-	void
-ROCAL_API_CALL rocalGetBoundingBoxCords(RocalContext p_context, float* buf)
+
+void
+ROCAL_API_CALL rocalGetBoundingBoxCords(RocalContext p_context, double* buf)
 {
     if (!p_context)
         THROW("Invalid rocal context passed to rocalGetBoundingBoxCords")
@@ -550,7 +551,7 @@ void ROCAL_API_CALL rocalBoxIOUMatcher(RocalContext p_context, std::vector<float
 }
 
 void
-ROCAL_API_CALL rocalCopyEncodedBoxesAndLables(RocalContext p_context, float* boxes_buf, int* labels_buf)
+ROCAL_API_CALL rocalCopyEncodedBoxesAndLables(RocalContext p_context, double* boxes_buf, int* labels_buf)
 {
     if (!p_context)
         THROW("Invalid rocal context passed to rocalCopyEncodedBoxesAndLables")
@@ -572,12 +573,12 @@ ROCAL_API_CALL rocalCopyEncodedBoxesAndLables(RocalContext p_context, float* box
         sum += meta_data.second->get_bb_labels_batch()[i].size();
     }
     // copy labels buffer & bboxes buffer parallely
-    #pragma omp parallel for
+#pragma omp parallel for
     for (unsigned i = 0; i < meta_data_batch_size; i++)
     {
         unsigned bb_count = meta_data.second->get_bb_labels_batch()[i].size();
         int *temp_labels_buf = labels_buf + bb_offset[i];
-        float *temp_bbox_buf = boxes_buf + (bb_offset[i] * 4);
+        double *temp_bbox_buf = boxes_buf + (bb_offset[i] * 4);
         memcpy(temp_labels_buf, meta_data.second->get_bb_labels_batch()[i].data(), sizeof(int) * bb_count);
         memcpy(temp_bbox_buf, meta_data.second->get_bb_cords_batch()[i].data(), sizeof(BoundingBoxCord) * bb_count);
     }
